pyproject.toml
[project]
name = "stock-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "fredapi>=0.5.2",
    "numpy>=2.4.3",
    "opendartreader>=0.2.3",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pykrx>=1.0.51",
    "pyportfolioopt>=1.6.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "yfinance>=1.2.0",
]
